requirements.txt
langchain>=0.3.25
langchain-core>=0.3.58
langchain-community>=0.0.21
pydantic>=2.7.4
openai>=1.14.3
streamlit>=1.32.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
matplotlib>=3.7.0
pandas>=2.0.0
msal==1.27.0
requests==2.31.0
google-auth==2.27.0
google-generativeai==0.3.2 